+++ src/pixelator/pna/sample_calling/sample_calling.py
@@ -31,7 +31,10 @@ logger = logging.getLogger(__name__)
 
 
 def collect_hash_info(
-    input_pxl_file: PNAPixelDataset, hashed_antibody_mapping: HashedAntibodyMapping
+    input_pxl_file: PNAPixelDataset,
+    hashed_antibody_mapping: HashedAntibodyMapping,
+    confidence_threshold: float,
+    undetermined_sample_name: str = "undetermined",
 ) -> pl.DataFrame:
     """Map components to samples in an sample-hashed dataset.
 
@@ -47,12 +50,14 @@ def collect_hash_info(
         input_pxl_file (PNAPixelDataset): The input pixel dataset.
         hashed_antibody_mapping (HashedAntibodyMapping): Mapping of sample names
             to hashed antibody names and the full list of hashing antibodies (from panel).
+        confidence_threshold (float): Confidence threshold.
+        undetermined_sample_name (str, optional): Name to use for undetermined components. Defaults to "undetermined".
 
     Returns:
         pl.DataFrame: A Polars DataFrame containing the following columns:
             - 'component': The component identifier.
             - '{sample}_hash_count': The summed hash count for each sample.
-            - 'undetermined_hash_count': The summed hash count for antibodies not mapped to any
+            - '{undetermined_sample_name}_hash_count': The summed hash count for antibodies not mapped to any
                 sample.
             - 'called_sample': The sample with the highest hash count for each component
                 (may be "undetermined").
@@ -63,9 +68,11 @@ def collect_hash_info(
     samples = hashed_antibody_mapping.keys()
     unmapped = hashed_antibody_mapping.unmapped_hashing_antibodies
     undetermined_col = (
-        pl.sum_horizontal(list(unmapped)).alias("undetermined_hash_count")
+        pl.sum_horizontal(list(unmapped)).alias(
+            f"{undetermined_sample_name}_hash_count"
+        )
         if unmapped
-        else pl.lit(0).alias("undetermined_hash_count")
+        else pl.lit(0).alias(f"{undetermined_sample_name}_hash_count")
     )
 
     hash_counts_per_sample = ab_count_data.select(
@@ -80,7 +87,7 @@ def collect_hash_info(
         undetermined_col,
     )
     unpivot_cols = [f"{sample}_hash_count" for sample in samples] + [
-        "undetermined_hash_count"
+        f"{undetermined_sample_name}_hash_count"
     ]
     called_sample = (
         hash_counts_per_sample.select(["component"] + unpivot_cols)
@@ -95,15 +102,32 @@ def collect_hash_info(
             pl.col("value").max().alias("max_value"),
             pl.col("value").sum().alias("total_value"),
         )
-    ).with_columns(sample_confidence=pl.col("max_value") / pl.col("total_value"))
+        .with_columns(
+            pl.when(pl.col("total_value") == 0)
+            .then(pl.lit(0.0))
+            .otherwise(pl.col("max_value") / pl.col("total_value"))
+            .alias("sample_confidence")
+        )
+        .with_columns(
+            pl.when(pl.col("sample_confidence") < confidence_threshold)
+            .then(pl.lit(undetermined_sample_name))
+            .otherwise(pl.col("called_sample"))
+            .alias("called_sample")
+        )
+    )
 
     result = hash_counts_per_sample.join(
         called_sample, on="component", how="left"
     ).select(
         ["component"]
         + [pl.col(f"{sample}_hash_count") for sample in samples]
-        + ["undetermined_hash_count", "called_sample", "sample_confidence"]
+        + [
+            f"{undetermined_sample_name}_hash_count",
+            "called_sample",
+            "sample_confidence",
+        ]
     )
+
     return result.collect()
 
 
@@ -248,35 +272,6 @@ def _build_post_sample_calling_anndata(
     return new_adata
 
 
-def _apply_confidence_threshold_to_hash_info(
-    hash_info: pl.DataFrame,
-    confidence_threshold: float,
-    undetermined_sample_name: str = "undetermined",
-) -> pl.DataFrame:
-    """Apply confidence threshold to hash info and assign undetermined components.
-
-    Components with confidence below the threshold are assigned to undetermined.
-
-    Args:
-        hash_info (pl.DataFrame): Hash info dataframe.
-        confidence_threshold (float): Confidence threshold.
-        undetermined_sample_name (str, optional): Name to use for undetermined components. Defaults to "undetermined".
-
-    Returns:
-        pl.DataFrame: Hash info dataframe where confidence levels have been applied.
-
-    """
-    return hash_info.with_columns(
-        pl.when(
-            (pl.col("sample_confidence") < confidence_threshold)
-            | (pl.col("called_sample") == "undetermined")
-        )
-        .then(pl.lit(undetermined_sample_name))
-        .otherwise(pl.col("called_sample"))
-        .alias("called_sample")
-    )
-
-
 def _find_nodes_to_remove(
     hashing_antibody_mapping: HashedAntibodyMapping,
     remove_incompatible: bool,
@@ -332,14 +327,13 @@ def sample_calling(
     Returns: List of all output pxl files created.
 
     """
-    hash_info = collect_hash_info(input_pxl, hashing_antibody_mapping)
-    panel = PNAAntibodyPanel.from_pxl_dataset(input_pxl)
-
-    hash_info = _apply_confidence_threshold_to_hash_info(
-        hash_info,
+    hash_info = collect_hash_info(
+        input_pxl,
+        hashing_antibody_mapping,
         confidence_threshold,
-        undetermined_sample_name=undetermined_sample_name,
+        undetermined_sample_name,
     )
+    panel = PNAAntibodyPanel.from_pxl_dataset(input_pxl)
 
     dehashed = hash_info.group_by("called_sample")
     output_files: list[Path] = []