+++ CHANGELOG.md
@@ -44,6 +44,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- sample calling will now use the pool name in the file name of undetermined
+  components to avoid file name collisions in the pipeline.
+
 ## [0.25.0] - 2026-04-15
 
 ### Added